Removed a bug that doesn't allow reset.
The bug was caused by a line of code that tested if the inputted string is a number or string.
Unfortunately the removal of this line of code means the reset function will not work accurately if the input value is a string.
A check for input value type will be added sometime later.